requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-barcode==0.14.0
redis==4.6.0
slowapi==0.1.8
boto3==1.28.40
botocore==1.31.40
bcrypt==4.0.1
email-validator==2.0.0.post2
python-multipart==0.0.6